requirements.txt
pywinauto
numpy
pandas
scikit-learn
pillow
scipy